refactor(tui): extract PaneState into state/pane.rs
Move the per-agent pane model — PaneState and its supporting enums
(PaneKind, Ownership, Autonomy, TailKind, PendingView) — into a
state/pane.rs submodule reached by qualified path. Methods and fields the
reducer touches from mod.rs (bucket, flush_tail, push, push_external,
stream, push_tool_diff, scroll_page_*, Autonomy::next, and the since /
last_bucket fields) are pub(super) — visible within the state module tree,
no wider. ui.rs and the outer tui/mod.rs import these from state::pane.

Shared consts (SCROLLBACK_CAP, TICKS_PER_SEC) and helpers (fmt_elapsed,
stop_label) stay at the state root; pane.rs reaches them via `super`. No
behavior change.
